move album CRUD hooks back to hooks folder
